Allow 10-bit PWM output (up to 16 bit)

+++ Src/GPIO/Communication/SerialGPIOFactory.cs
@@ -118,7 +118,7 @@ namespace R2Core.GPIO
 
 		}
 
-        public IOutputPort<byte> CreateAnalogOutput(string id, int port, int nodeId = ArduinoSerialPackageFactory.DEVICE_NODE_LOCAL) {
+        public IOutputPort<int> CreateAnalogOutput(string id, int port, int nodeId = ArduinoSerialPackageFactory.DEVICE_NODE_LOCAL) {
 
             return InstantiateDevice(new SerialAnalogOutput(id, m_devices.GetNode(nodeId), m_connection, port));
